refactor: Extract methods and classes to separate files
References: https://outsystemsrd.atlassian.net/browse/RMET-2991
